+++ addons/point_of_sale/static/src/app/base_models/base_orderline.js
@@ -1,0 +1,5 @@
+/* @odoo-module */
+
+import { PosModel } from "@point_of_sale/app/base_models/base";
+
+export class Orderline extends PosModel {}
